Django-Project/school_admission/requirements.txt
Adafruit-Blinka==8.19.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-requests==1.14.0
adafruit-circuitpython-typing==1.9.4
Adafruit-PlatformDetect==3.46.0
Adafruit-PureIO==1.1.11
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bidict==0.22.1
bootstrap4==0.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
crispy-bootstrap5==0.7
cycler==0.11.0
distlib==0.3.6
Django==4.2.2
django-crispy-forms==2.0
django-js-asset==2.1.0
example-package==1.0.0
fastapi==0.101.1
fastapi-socketio==0.0.10
filelock==3.12.0
fonttools==4.40.0
frozenlist==1.4.0
h11==0.14.0
htmx==0.0.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
markdown2==2.4.10
MarkupSafe==2.1.3
matplotlib==3.7.1
multidict==6.0.4
mysqlclient==2.1.1
nicegui==1.3.9
numpy==1.25.0rc1
orjson==3.9.5
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.1
pscript==0.7.7
pydantic==1.10.12
pyftdi==0.54.0
Pygments==2.16.1
pyparsing==3.1.0b2
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-multipart==0.0.6
python-socketio==5.8.0
pyusb==1.2.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.6.3
tzdata==2023.3
uvicorn==0.22.0
vbuild==0.8.2
virtualenv==20.23.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
